+++ packages/generator-ts-project/src/Components/Inquiry/Prompts/QuestionSetProperty.ts
@@ -1,0 +1,13 @@
+/**
+ * Represents a dynamic proparty for a question-set.
+ *
+ * @template T
+ * The type of the property.
+ *
+ * @template TResult
+ * The type of the answer-hash of the inner questions.
+ *
+ * @template TAnswers
+ * The type of the answers.
+ */
+export type QuestionSetProperty<T, TResult, TAnswers> = T | ((result: TResult, answers?: TAnswers) => (T | Promise<T>));