+++ CHANGELOG.md
@@ -6,7 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## TSProjectGenerator [Unreleased]
 
-[Show differences](https://github.com/manuth/TSProjectGenerator/compare/v3.0.4...dev)
+[Show differences](https://github.com/manuth/TSProjectGenerator/compare/v3.0.5...dev)
+
+## TSProjectGenerator v3.0.5
+### Fixed
+  - The `PathPrompt`s treatment of trailing backslashes on linux
+  - Broken unit-tests
+
+### Added
+  - Support for getting all answers from within a `QuestionSetPrompt`
+
+[Show differences](https://github.com/manuth/TSProjectGenerator/compare/v3.0.4...v3.0.5)
 
 ## TSProjectGenerator v3.0.4
 ### Added